Initial implementation of SHA256
Adding a 'common.c' file for now. I am considering having an optional #define to use 'common.c' instead of header-specific methods in the event that it isn't a big deal to include multiple files. Having the option to only need to include one file (stb-style) is still a goal for this library.