Add message remainder to buffer before passing to 'SHA2_ApplyPadding()'
This change makes the main hashing functions responsible for loading the message remainder into the buffer and SHA2_ApplyPadding() only responsible for applying the actual padding in preparation for implementing file hashing for SHA2. Hashing strings and hashing files use different sources of input and 'SHA2_ApplyPadding()' shouldn't be responsible for dealing with the message remainder.