Batch file for building test driver returns with error code on failure
This change modifies the test driver build batch file to pass on the error code it receives from MSVC if the test build fails. This prevents the test driver output from drowning out errors when executing the Sublime Text build system and the build fails. The build system was already checking for an error code before attempting to run the test driver but the batch file was always returning success.